+++ app/src/Component/Offering/QualifiedInstructor/QualifiedInstructorTable.tsx
@@ -1,4 +1,4 @@
-import { ITableWrapperProps } from "../offeringUtils"
+import { ITableWrapperProps } from "~/Component/Offering"
 import { Row, Col } from "antd"
 import React from "react"
 import moment from "moment"
